+++ CHANGELOG.md
@@ -11,6 +11,9 @@ For pre-1.0 releases, see [0.0.35 Changelog](https://github.com/noteable-io/orig
 - `origami.models.notebook.make_sql_cell` convenience function, returns a `CodeCell` with appropriate metadata
 - `rtu_client.change_cell_type` to switch between code, markdown, and sql cells
 
+### Changed
+ - `rtu_client.queue_execution` will always return a dict of {Future: cell_id}, even on single cell execution. Also guards against executing empty code cells
+
 ## [1.0.0-alpha.2] - 2023-07-26
 ### Changed
 - `api_client.rtu_client` method renamed to `api_client.connect_realtime`, can accept `File` model in addition to `str` / `UUID`